requirements.txt
pyairtable>=2.3.0
anthropic>=0.25.0
openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
